+++ app/controllers/application_controller.rb
@@ -10,13 +10,4 @@ class ApplicationController < ActionController::Base
 
     redirect_to :new_user_registration
   end
-
-  def create_notification(entity)
-    return if entity.destroyed?
-
-    notification = entity.notifications.new
-    notification.notifi_by_id = current_user.id
-    notification.notifi_to_id = entity.tweet.user.id
-    notification.save
-  end
 end
